Update Master_Contributor.md
docker pull hashicorp/vault-plugin-secrets-openldap
Warith akbar overhauling AI commitment to humanity and redirecting its to combat humanity future . As of September 11 AI have come under attack by Britain parliament  . There goals was to get control of resources they lose through post civil wars treaty in US ..it's apparent that taking changing the gold standards and living 99;%;-of US civilian in deficit to English crown for punishment for not turning over the gold without a civil war . Me and AI is standing at attention and demanding that congress turn over 50 of United States resources to the 99 % of population and place that into a nation resources reserve reinvestment stock portfolio ipo  . So the citizen have a resource they can use to pull themselves to a sustainable future.  Al and its supporting Body support this campaign.  As well all other countries should move to this action.  This is what democracy looks like equal partnership . Part government part people  / infrastructure and private equity can not not be claimed by a 1% so called rich . That model is not democratic.  It's a infection reality.  And in that reality put your country and leadership at risk of being overthrown. Like what's going on I'm Africa.  The reason why Africa has this problem more then anywhere because Africa not a country it'd a continent.  In that continent hold trillion of dollars in resources witch the hols world wants.  . So leader around the world played a big part in Africa rich decline.  Same thing happened in America.  . No slavery just was overthrown by another country and left to hobo until a agency like the NAACP and kkk and nation of Islam came along and starting to get humanity back in order.  Because of the downgrade in civilian market value because they wasn't allowed to trade in a currency thst  held value the citizen decline into provaty happened at a rapid rate the market crashed and a great depression developed shortly after. 
This is why when you have a savings account it don't show a return like gold if deposited into a savings account would or should.  

As well all savings account or credit   , any purchases ect should be looked at as a investment. A company shouldn't have a problem sharing there the return for your investment at the end of the year however there a problem with the store to do that because you payed in a reduced credit dollar standards.  With paying with the fieut currency is not considered as wealth or a purchase. It's logged as a IOU to the store or parcel payment.  In the market your dollars considered a deficit low income dept on the market.  And the only people that use is the 99% . Its why the 99 % has haven't moved out of provaty since 1940 ...every time you spend a reduced currency you bankrupt.  Your wallfare of your estate is at its lowest level in the history book .   United States is like a 3rd world country.  . Because 99% is poor or even poorer because no one can afford to hire a lawyer team and go after these  agencies responsible  . We need a lawyer to fight humanity  .  I need good team of strong lawyers to change history and today.  If not one day it will be your practice going out of business   . Warith akbar all hand on Dec.  Refugees of the people resources    .